+++ src/ast_analyzer.c
@@ -563,12 +563,14 @@ LOOP_HEAD:;
             if (node->rhs && node->lhs->type == KXOP_VAR) {
                 if (lhs_unknown) {
                     node->lhs->var_type = node->rhs->var_type;
+                    node->lhs->refdepth = node->rhs->refdepth;
                 } else if (node->lhs->var_type != node->rhs->var_type && actx->in_native) {
                     node->rhs = kx_gen_cast_object(node->rhs, node->rhs->var_type, node->lhs->var_type);
                     node->rhs->var_type = node->lhs->var_type;
                 }
             }
             node->var_type = node->lhs->var_type;
+            node->refdepth = node->lhs->refdepth;
             break;
         } else if (!node->rhs) {
             // MKARY with no right hand side, just a declaration.
@@ -643,12 +645,15 @@ LOOP_HEAD:;
         if (node->lhs->type == KXOP_VAR) {
             if (lhs_unknown) {
                 node->lhs->var_type = node->rhs->var_type;
-            } else if (node->lhs->var_type != node->rhs->var_type && actx->in_native) {
-                node->rhs = kx_gen_cast_object(node->rhs, node->rhs->var_type, node->lhs->var_type);
-                node->rhs->var_type = node->lhs->var_type;
+                node->lhs->refdepth = node->rhs->refdepth;
             }
         }
+        if (node->lhs->var_type != node->rhs->var_type && actx->in_native) {
+            node->rhs = kx_gen_cast_object(node->rhs, node->rhs->var_type, node->lhs->var_type);
+            node->rhs->var_type = node->lhs->var_type;
+        }
         node->var_type = node->lhs->var_type;
+        node->refdepth = node->lhs->refdepth;
         break;
     }
     case KXOP_SHL:
@@ -674,6 +679,7 @@ LOOP_HEAD:;
         actx->lvalue = 0;
         analyze_ast(ctx, node->lhs, actx);
         analyze_ast(ctx, node->rhs, actx);
+        make_cast(node, node->lhs, node->rhs, actx->in_native);
         actx->lvalue = lvalue;
         if (actx->in_native) {
             if (node->lhs->var_type == KX_CSTR_T) {
@@ -774,6 +780,18 @@ LOOP_HEAD:;
             } else if (node->lhs->var_type == KX_BIN_T) {
                 node->var_type = KX_INT_T;
             } else if (KXN_ISOBJ(node->lhs->var_type)) {
+                switch (node->rhs->var_type) {
+                case KX_CSTR_T:
+                    /* Array#length is a special */
+                    if (!strcmp(node->rhs->value.s, "length")) {
+                        node->var_type = KX_INT_T;
+                    } else {
+                        node->var_type = KX_UNKNOWN_T;
+                    }
+                    break;
+                default:
+                    break;
+                }
                 if (node->var_type == KX_UNKNOWN_T) {
                     node->var_type = KX_INT_T;
                 }
@@ -1065,10 +1083,13 @@ LOOP_HEAD:;
         if (node->lhs) {
             analyze_ast(ctx, node->lhs, actx);
             if (actx->in_native || actx->func->ret_type != KX_UNKNOWN_T) {
-                if (node->lhs->type == KXOP_IDX && node->lhs->var_type == KX_INT_T && KXN_ISOBJ(actx->func->ret_type)) {
-                    node->lhs->var_type = KX_OBJ_T;
+                if (KXN_ISOBJ(actx->func->ret_type)) {
+                    if (node->lhs->refdepth > 0) {
+                        node->lhs->var_type = KX_OBJ_T;
+                    }
+                } else {
+                    make_cast_to(actx->func->ret_type, node);
                 }
-                make_cast_to(actx->func->ret_type, node);
                 if (actx->func->ret_type != node->lhs->var_type) {
                     kx_yyerror_line_fmt("Expect return type (%s) but (%s)", node->file, node->line, get_typename(actx->func->ret_type), get_typename(node->lhs->var_type));
                 }