+++ ChangeLog.md
@@ -10,7 +10,6 @@
 
 *   Planned
     *   Support to read/write from/to Clipboard.
-    *   Support Array of Double in native.
     *   Support Big Integer in native.
     *   User selectable installation.
     *   Documentation & SpecTest complete.
@@ -19,6 +18,7 @@
 ## V0.16.0 (Current Development Version)
 
 *   Updated
+    *   Support Array of Double in native.
     *   updated some SpecTest pages.
 
 ## V0.15.3 (15th-3 Preview Release) - 2020/11/27